override templates with same schema from projects
