+++ packages/core/src/schema/configFiles/template.ts
@@ -1,54 +1,11 @@
 import { z } from 'zod';
 import { DefaultsSchema, EnginesSchema } from './project';
-
-const NpmDep = z.record(z.string(), z.string());
-
-const BootstrapQuestionsSchema = z
-    .array(
-        z.object({
-            options: z
-                .array(
-                    z.object({
-                        title: z.string(),
-                        value: z.object({}),
-                    })
-                )
-                .optional(),
-            configProp: z
-                .object({
-                    prop: z.string(),
-                    key: z.string(),
-                })
-                .optional(),
-            type: z.string(),
-            title: z.string(),
-        })
-    )
-    .describe('Defines list of custom bootstrap questions');
+import { TemplateConfig } from '../shared';
 
 export const RootTemplateSchema = z.object({
     defaults: z.optional(DefaultsSchema),
     engines: z.optional(EnginesSchema),
-    templateConfig: z
-        .object({
-            includedPaths: z
-                .array(z.string())
-                .describe('Defines list of all file/dir paths you want to include in template')
-                .optional(),
-            bootstrapQuestions: BootstrapQuestionsSchema,
-            packageTemplate: z.optional(
-                z.object({
-                    dependencies: z.optional(NpmDep),
-                    devDependencies: z.optional(NpmDep),
-                    peerDependencies: z.optional(NpmDep),
-                    optionalDependencies: z.optional(NpmDep),
-                    name: z.string().optional(),
-                    version: z.string().optional(),
-                })
-            ),
-        })
-        .describe('Used in `renative.template.json` allows you to define template behaviour.')
-        .optional(),
+    templateConfig: TemplateConfig.optional(),
 });
 
 // {
