+++ packages/core/src/schema/shared/index.ts
@@ -41,3 +41,48 @@ export const BuildSchemeFragment = {
             )
     ),
 };
+
+const NpmDep = z.record(z.string(), z.string());
+
+const BootstrapQuestionsSchema = z
+    .array(
+        z.object({
+            options: z
+                .array(
+                    z.object({
+                        title: z.string(),
+                        value: z.object({}),
+                    })
+                )
+                .optional(),
+            configProp: z
+                .object({
+                    prop: z.string(),
+                    key: z.string(),
+                })
+                .optional(),
+            type: z.string(),
+            title: z.string(),
+        })
+    )
+    .describe('Defines list of custom bootstrap questions');
+
+export const TemplateConfig = z
+    .object({
+        includedPaths: z
+            .array(z.string())
+            .describe('Defines list of all file/dir paths you want to include in template')
+            .optional(),
+        bootstrapQuestions: BootstrapQuestionsSchema.optional(),
+        packageTemplate: z.optional(
+            z.object({
+                dependencies: z.optional(NpmDep),
+                devDependencies: z.optional(NpmDep),
+                peerDependencies: z.optional(NpmDep),
+                optionalDependencies: z.optional(NpmDep),
+                name: z.string().optional(),
+                version: z.string().optional(),
+            })
+        ),
+    })
+    .describe('Used in `renative.template.json` allows you to define template behaviour.');
